Add BaseOptimizer
- move general optimization methods from Nsgaii to BaseOptimizer
