+++ modules/nsgaii.py
@@ -22,34 +22,26 @@ __author__ = "Heta Rekilä \n Juhani Sundell \n Tuomas Pitkänen"
 __version__ = "2.0"
 
 import collections
-import math
 import os
 import subprocess
 from pathlib import Path
 from timeit import default_timer as timer
 
 import numpy as np
-import rx
-from rx import operators as ops
 
-from . import file_paths as fp
 from . import general_functions as gf
 from . import math_functions as mf
 from . import optimization as opt
 from .concurrency import CancellationToken
 from .element_simulation import ElementSimulation
-from .energy_spectrum import EnergySpectrum
 from .enums import IonDivision
 from .enums import OptimizationState
 from .enums import OptimizationType
-from .mcerd import MCERD
-from .observing import Observable
-from .parsing import CSVParser
 from .point import Point
 from .recoil_element import RecoilElement
 
 
-class Nsgaii(Observable):
+class Nsgaii(opt.BaseOptimizer):
     """
     Class that handles the NSGA-II optimization. This needs to handle both
     fluence and recoil element optimization. Recoil element optimization
@@ -71,60 +63,46 @@ class Nsgaii(Observable):
         """
         Initialize the NSGA-II algorithm with needed parameters and start
         running it.
+
+        Only NSGA-II-specific arguments are documented here. See
+        BaseOptimizer for general arguments.
+
         Args:
             gen: Number of generations to be done.
-            element_simulation: ElementSimulation object that is optimized.
             pop_size: Population size.
             sol_size: Amount of variables in one solution.
-            upper_limits: Upper limit(s) for variables in a solution.
-            lower_limits: Lower limit(s) for a variable in a solution.
-            optimization_type: Whether to optimize recoil or fluence.
-            recoil_type: Type of recoil: either "box" (4 points or 5),
-                "two-peak" (high areas at both ends of recoil, low in the
-                middle) or "free" (no limits to the shape of the recoil).
-                number_of_processes: How many processes are used in MCERD
-                calculation.
             cross_p: Crossover probability.
             mut_p: Mutation probability, should be something small.
-            stop_percent: When to stop running MCERD (based on the ratio in
-            average change between checkups).
-            check_time: Time interval for checking if MCERD should be stopped.
-            ch: Channel with for running get_espe.
-                used in comparing the simulated energy spectra.
             dis_c: Distribution index for crossover. When this is big,
                 a new solution is close to its parents.
             dis_m: Distribution for mutation.
-            check_max: Maximum time for running a simulation.
-            check_min: Minimum time for running a simulation.
-            skip_simulation: whether simulation is skipped altogether
-            use_efficiency: whether to use efficiency for pre-calculated
-                spectrum.
         """
         # TODO separate the two optimization types into two classes
-        Observable.__init__(self)
+        # Observable.__init__(self)
+        opt.BaseOptimizer.__init__(
+            self,
+            element_simulation=element_simulation,
+            upper_limits=upper_limits,
+            lower_limits=lower_limits,
+            optimization_type=optimization_type,
+            recoil_type=recoil_type,
+            number_of_processes=number_of_processes,
+            stop_percent=stop_percent,
+            check_time=check_time,
+            ch=ch,
+            measurement=measurement,
+            cut_file=cut_file,
+            check_max=check_max,
+            check_min=check_min,
+            skip_simulation=skip_simulation,
+            use_efficiency=use_efficiency,
+            verbose=verbose,
+            optimize_by_area=optimize_by_area
+        )
+
         self.evaluations = gen * pop_size
-        self.element_simulation = element_simulation  # Holds other needed
-        # information including recoil points and access to simulation settings
         self.pop_size = pop_size
         self.sol_size = sol_size
-        self.upper_limits = upper_limits
-        if not self.upper_limits:
-            self.upper_limits = [120, 1]
-        self.lower_limits = lower_limits
-        if self.lower_limits is None:
-            self.lower_limits = [0.01, 0.0001]
-        self.optimization_type = optimization_type
-        self.rec_type = recoil_type
-
-        # MCERD-specific parameters
-        self.number_of_processes = number_of_processes
-        self._skip_simulation = skip_simulation
-        self.stop_percent = stop_percent
-        self.check_time = check_time
-        self.check_max = check_max
-        self.check_min = check_min
-
-        self.channel_width = ch
 
         # Crossover and mutation parameters
         self.cross_p = cross_p
@@ -135,14 +113,7 @@ class Nsgaii(Observable):
         self.bit_length_x = 0
         self.bit_length_y = 0
 
-        self.measurement = measurement
-        self.cut_file = Path(cut_file)
-
         self.population = None
-        self.measured_espe = None
-        self.use_efficiency = use_efficiency
-        self.verbose = verbose
-        self.optimize_by_area = optimize_by_area
 
     def _prepare_optimization(self, initial_pop=None,
                               cancellation_token=None,
@@ -184,104 +155,6 @@ class Nsgaii(Observable):
             self.run_initial_simulation(cancellation_token, ion_division)
 
         self.population = self.evaluate_solutions(initial_pop)
-
-    def prepare_measured_spectra(self):
-        """Calculate measured spectra and parse it.
-
-        Optimized solutions are compared to the measured spectra.
-        """
-        EnergySpectrum.calculate_measured_spectra(
-            self.measurement, [self.cut_file], self.channel_width,
-            no_foil=True, use_efficiency=self.use_efficiency)
-
-        # TODO maybe just use the value returned by calc_spectrum?
-        # Add result files
-        hist_file = Path(self.measurement.get_energy_spectra_dir(),
-                         f"{self.cut_file.stem}.no_foil.hist")
-
-        parser = CSVParser((0, float), (1, float))
-        self.measured_espe = list(parser.parse_file(hist_file, method="row"))
-
-    def combine_previous_erd_files(self):
-        """Combine previous erd files to used as the starting point."""
-        erd_file_name = fp.get_erd_file_name(
-            self.element_simulation.get_main_recoil(), "combined",
-            optim_mode=self.optimization_type)
-
-        gf.combine_files(self.element_simulation.get_erd_files(),
-                         Path(self.element_simulation.directory,
-                              erd_file_name))
-
-    def run_initial_simulation(self, cancellation_token, ion_division):
-        """Run initial element simulation.
-        """
-        def stop_if_cancelled(
-                optim_ct: CancellationToken, mcerd_ct: CancellationToken):
-            optim_ct.stop_if_cancelled(mcerd_ct)
-            return mcerd_ct.is_cancellation_requested()
-
-        ct = CancellationToken()
-        observable = self.element_simulation.start(
-            self.number_of_processes, start_value=201,
-            optimization_type=self.optimization_type,
-            ct=ct, print_output=True, max_time=self.check_max,
-            ion_division=ion_division)
-
-        if observable is not None:
-            self.on_next(self._get_message(
-                OptimizationState.SIMULATING,
-                evaluations_left=self.evaluations))
-
-            ct_check = rx.timer(0, 0.2).pipe(
-                ops.take_while(lambda _: not stop_if_cancelled(
-                    cancellation_token, ct)),
-                ops.filter(lambda _: False),
-            )
-            # FIXME spectra_chk should only be performed when pre-simulation
-            #   has finished, otherwise there will be no new observed atoms
-            #   and the difference between the two spectra is 0
-            spectra_chk = rx.timer(self.check_min, self.check_time).pipe(
-                ops.merge(ct_check),
-                ops.map(lambda _: get_optim_espe(
-                    self.element_simulation, self.optimization_type)),
-                ops.scan(
-                    lambda prev_espe, next_espe: (prev_espe[1], next_espe),
-                    seed=[None, None]),
-                ops.map(lambda espes: calculate_change(
-                    *espes, self.element_simulation.channel_width)),
-                ops.take_while(
-                    lambda change: change > self.stop_percent and not
-                    ct.is_cancellation_requested()
-                ),
-                ops.do_action(
-                    on_completed=ct.request_cancellation)
-            )
-            merged = rx.merge(observable, spectra_chk).pipe(
-                ops.take_while(
-                    lambda x: not isinstance(x, dict) or x[
-                        MCERD.IS_RUNNING],
-                    inclusive=True)
-            )
-            # Simulation needs to finish before optimization can start
-            # so we run this synchronously.
-            # TODO use callback instead of running sync
-            merged.run()
-            # TODO should not have to call this manually
-            self.element_simulation._clean_up(ct)
-
-        else:
-            raise ValueError(
-                "Could not start simulation. Check that simulation is not "
-                "currently running.")
-
-    @staticmethod
-    def _get_message(state, **kwargs):
-        """Returns a dictionary with the state of the optimization and
-        other """
-        return {
-            "state": state,
-            **kwargs
-        }
 
     @staticmethod
     def crowding_distance(front_no, objective_values):
@@ -432,6 +305,8 @@ class Nsgaii(Observable):
         self.bit_length_x = len_of_x
         self.bit_length_y = len_of_y
 
+    # TODO: Move to BaseOptimizer (deal with self.sol_size)
+    # TODO: Reduce repetition
     def form_recoil(self, current_solution, name=""):
         """
         Form recoil based on solution size.
@@ -796,33 +671,6 @@ class Nsgaii(Observable):
                         + self.lower_limits
 
         return init_sols
-
-    def modify_measurement(self):
-        """
-        Modify measured energy spectrum to match the simulated in regards to
-        the x coordinates.
-        """
-        new = []
-        i = 0
-        # Add zero points to start and end to get correct mean values
-        first_x = self.measured_espe[0][0]
-        last_x = self.measured_espe[-1][0]
-
-        # TODO could use deque for quicker inserts
-        self.measured_espe.insert(
-            0, (round(first_x - self.element_simulation.channel_width, 4), 0.0))
-        self.measured_espe.append(
-            (round(last_x + self.element_simulation.channel_width, 4), 0.0))
-
-        while i < len(self.measured_espe) - 1:  # Do nothing to the last point
-            current_point = self.measured_espe[i]
-            next_point = self.measured_espe[i + 1]
-
-            new_x = round((next_point[0] + current_point[0]) / 2, 4)
-            new_y = round((next_point[1] + current_point[1]) / 2, 5)
-            new.append((new_x, new_y))
-            i += 1
-        self.measured_espe = new
 
     @staticmethod
     def nd_sort(pop_obj, n, r_n=np.inf):
@@ -1411,36 +1259,3 @@ def get_ys(y_lower, y_upper, pop_size, z=None, lower_limit_at_first=False):
             return np.array([low_limit, y_coords[0], y_coords[1],
                              y_coords[2], low_limit])
     return np.array([y_coords[0], y_coords[1], y_coords[2], low_limit])
-
-
-def get_optim_espe(elem_sim: ElementSimulation,
-                   optimization_type: OptimizationType):
-    if optimization_type is OptimizationType.RECOIL:
-        recoil = elem_sim.optimization_recoils[0]
-    else:
-        recoil = elem_sim.get_main_recoil()
-
-    espe, _ = elem_sim.calculate_espe(
-        recoil, optimization_type=optimization_type, write_to_file=False)
-    return espe
-
-
-def calculate_change(espe1, espe2, channel_width):
-    if not espe1 or not espe2:
-        return math.inf
-    uniespe1, uniespe2 = gf.uniform_espe_lists(
-        espe1, espe2, channel_width=channel_width)
-
-    # Calculate distance between energy spectra
-    # TODO move this to math_functions
-    sum_diff = 0
-    amount = 0
-    for point1, point2 in zip(uniespe1, uniespe2):
-        if point1[1] != 0 or point2[1] != 0:
-            amount += 1
-            sum_diff += abs(point1[1] - point2[1])
-    # Take average of sum_diff (non-zero diffs)
-    if amount:
-        return sum_diff / amount
-    else:
-        return math.inf
